refactor credential form pages into components
